provides presentation

+++ domain/src/main/java/com/tailspin/domain/model/Item.kt
@@ -3,7 +3,7 @@ package com.tailspin.domain.model
 data class Item (
     var id : Int,
     var name : String,
-    var time : Int,
+    var time : Long,
     var tags : List<Tag>,
     var amount : Int
 )
